Clarified that `conj` can also be augmented in the enhanced graph.
This was part of the original proposal in Schuster and Manning (2016) but it was mistakenly omitted from the v2 guidelines. Nevertheless, the Stanford Enhancer produces it and it was always part of enhanced UD in some languages (English, Swedish, Dutch). So the core group now decided that it should be explicitly mentioned in the guidelines.